requirements.txt
awscli==1.16.162
botocore==1.12.152
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
credentials==1.1
dj-database-url==0.5.0
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
docutils==0.14
firebase-admin==2.16.0
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
msgpack==0.6.1
oauthlib==3.1.0
protobuf==3.7.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.2.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
views==0.3
virtualenv==16.3.0
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
